Changes to support beamsets when beamforming.
Beamsets allow for different antennas weighting schemes to be applied
on a beam-by-beam basis. In order to support this it is necessary to
for the following:
- parse out the antennas weightings for each beam from the delay model
- create a set of these weightings
- for each set create scales and offsets for the coherent and
  incoherent beams
- for each set create an incoherent beam that will be used for
  subtraction in the coherent beamformer
- apply the correct scales and weights in both the incoherent and
  coherent beamformers
- subtract the correct incoherent beam in the coherent beamformer
  based on a mapping of beam to beamset

This commit contains only compiling code and is likely not working yet.

Extensive changes to the testing are required to ensure that the logic
behind supporting multiple beamsets is robust (some of these changes are
included here).